Added some commented logging statements, fixing cluster indirect values

+++ ESzabbix.py
@@ -43,10 +43,13 @@ clustercache_file = "/tmp/clusterstats.cache." + user
 nodescache_file = "/tmp/nodestats.cache." + user
 lock_file="/tmp/ESzabbix.lock." + user
 
+# now = time.time()
+# logging = open("/tmp/whathappens.log."+str(now)+"-"+sys.argv[1]+"-"+sys.argv[2], "w");
+# logging.write(" ".join(sys.argv)+"\n")
 # Waiting to somebody write the cache
 while os.access(lock_file,os.F_OK):
+#    logging.write("Waiting a second ...\n")
     time.sleep(1)
-
 
 # __main__
 
@@ -63,12 +66,15 @@ try:
 except Exception, e:
     zbx_fail()
 
-if sys.argv[1] == 'cluster':
+if sys.argv[1] == 'cluster' and sys.argv[2] in clusterkeys_direct:
     nodestats = None
+#    now=time.strftime("%Y%m%d-%H:%M:%S")
     if use_cache(clustercache_file):
+#	logging.write(str(now) + ": Using cluster cache\n")
         nodestats = shelve.open(clustercache_file)
         nodestats = nodestats['stats']
     else:
+#	logging.write(str(now) + ": Generate lockfile and cluster cache\n")
         lock=open (lock_file, "w")
         try:
             nodestats = conn.cluster.stats()
@@ -80,73 +86,21 @@ if sys.argv[1] == 'cluster':
             if os.access(lock_file, os.F_OK):
                 os.remove(lock_file)
             zbx_fail()
-        os.remove(lock_file)
-    if sys.argv[2] in clusterkeys_direct:
-        if sys.argv[2] in docskeys:
-               returnval = nodestats['indices']['docs'][sys.argv[2]]
-        elif sys.argv[2] in storekeys:
-               returnval = nodestats['indices']['store'][sys.argv[2]]
-    elif sys.argv[2] in clusterkeys_indirect:
-        subtotal = 0
-        for nodename in conn.nodes.info()['nodes']:
-            try:
-		    if sys.argv[2] in indexingkeys:
-			indexstats = nodestats['nodes'][nodename]['indices']['indexing']
-		    elif sys.argv[2] in getkeys:
-			indexstats = nodestats['nodes'][nodename]['indices']['get']
-		    elif sys.argv[2] in searchkeys:
-			indexstats = nodestats['nodes'][nodename]['indices']['search']
-            except Exception, e:
-                pass
-            try:
-		if sys.argv[2] in indexstats:
-			subtotal += indexstats[sys.argv[2]]
-            except Exception, e:
-                pass
-        returnval = subtotal
-
-    else:
-        # Try to pull the managers object data
-        try:
-            escluster = conn.cluster
-        except Exception, e:
-            if sys.argv[2] == 'status':
-                returnval = "red"
-            else:
-                zbx_fail()
-        # Try to get a value to match the key provided
-        try:
-            returnval = escluster.health()[sys.argv[2]]
-        except Exception, e:
-            if sys.argv[2] == 'status':
-                returnval = "red"
-            else:
-                zbx_fail()
-        # If the key is "status" then we need to map that to an integer
-        if sys.argv[2] == 'status':
-            if returnval == 'green':
-                returnval = 0
-            elif returnval == 'yellow':
-                returnval = 1
-            elif returnval == 'red':
-                returnval = 2
-            else:
-                zbx_fail()
-
-# Mod to check if ES service is up
-elif sys.argv[1] == 'service':
-    if sys.argv[2] == 'status':
-        if conn.ping():
-            returnval = 1
-        else:
-            returnval = 0
-
-else: # Not clusterwide, check the next arg
+        if os.access(lock_file, os.F_OK):
+            os.remove(lock_file)
+    if sys.argv[2] in docskeys:
+        returnval = nodestats['indices']['docs'][sys.argv[2]]
+    elif sys.argv[2] in storekeys:
+        returnval = nodestats['indices']['store'][sys.argv[2]]
+elif sys.argv[1] == 'cluster' and  sys.argv[2] in clusterkeys_indirect:
     nodestats = None
+#    now=time.strftime("%Y%m%d-%H:%M:%S")
     if use_cache(nodescache_file):
+#	logging.write(str(now)+": Using node cache\n")
         nodestats = shelve.open(nodescache_file)
         nodestats = nodestats['stats']
     else:
+#	logging.write(str(now)+": Generate lockfile and node cache\n")
         lock=open (lock_file, "w")
         try:
             nodestats = conn.nodes.stats()
@@ -158,9 +112,88 @@ else: # Not clusterwide, check the next arg
             if os.access(lock_file, os.F_OK):
                 os.remove(lock_file)
             zbx_fail()
-        os.remove(lock_file)
+        if os.access(lock_file, os.F_OK):
+            os.remove(lock_file)
+        else:
+            pass # something is wrong if this is executed ...
+    subtotal = 0
+    for nodename in conn.nodes.info()['nodes'].keys():
+        try:
+    	    if sys.argv[2] in indexingkeys:
+    		    indexstats = nodestats['nodes'][nodename]['indices']['indexing']
+    	    elif sys.argv[2] in getkeys:
+    		    indexstats = nodestats['nodes'][nodename]['indices']['get']
+    	    elif sys.argv[2] in searchkeys:
+    		    indexstats = nodestats['nodes'][nodename]['indices']['search']
+        except Exception, e:
+            pass
+        try:
+    	    if sys.argv[2] in indexstats:
+    		    subtotal += indexstats[sys.argv[2]]
+        except Exception, e:
+            pass
+    returnval = subtotal
 
-    for nodename in conn.nodes.info()['nodes']:
+elif sys.argv[1] == "cluster":
+    # Try to pull the managers object data
+    try:
+        escluster = conn.cluster
+    except Exception, e:
+        if sys.argv[2] == 'status':
+            returnval = "red"
+        else:
+            zbx_fail()
+    # Try to get a value to match the key provided
+    try:
+        returnval = escluster.health()[sys.argv[2]]
+    except Exception, e:
+        if sys.argv[2] == 'status':
+            returnval = "red"
+        else:
+            zbx_fail()
+    # If the key is "status" then we need to map that to an integer
+    if sys.argv[2] == 'status':
+        if returnval == 'green':
+            returnval = 0
+        elif returnval == 'yellow':
+            returnval = 1
+        elif returnval == 'red':
+            returnval = 2
+        else:
+            zbx_fail()
+
+# Mod to check if ES service is up
+elif sys.argv[1] == 'service':
+    if sys.argv[2] == 'status':
+        if conn.ping():
+            returnval = 1
+        else:
+            returnval = 0
+
+else: # Not clusterwide, check the next arg
+    nodestats = None
+#    now=time.strftime("%Y%m%d-%H:%M:%S")
+    if use_cache(nodescache_file):
+#	logging.write(str(now)+": Usando cache nodos\n")
+        nodestats = shelve.open(nodescache_file)
+        nodestats = nodestats['stats']
+    else:
+#	logging.write(str(now)+": Creando lockfile y cache nodo\n")
+        lock=open (lock_file, "w")
+        try:
+            nodestats = conn.nodes.stats()
+            shelf = shelve.open(nodescache_file)
+            shelf['stats']=nodestats
+            shelf.close()
+            lock.close()
+        except Exception, e:
+            if os.access(lock_file, os.F_OK):
+                os.remove(lock_file)
+            zbx_fail()
+        if os.access(lock_file, os.F_OK):
+            os.remove(lock_file)
+
+    for nodename in conn.nodes.info()['nodes'].keys():
         if sys.argv[1] in nodestats['nodes'][nodename]['name']:
             if sys.argv[2] in indexingkeys:
                 stats = nodestats['nodes'][nodename]['indices']['indexing']
@@ -177,6 +210,9 @@ else: # Not clusterwide, check the next arg
             except Exception, e:
                 pass
 
+# now=time.strftime("%Y%m%d-%H:%M:%S")
+# logging.write(str(now)+": Finalizando ("+str(returnval)+")\n")
+# logging.close()
 
 # If we somehow did not get a value here, that's a problem.  Send back the standard 
 # ZBX_NOTSUPPORTED
@@ -184,5 +220,4 @@ if returnval is None:
     zbx_fail()
 else:
     print returnval
-
 # End
